Skip heuristics-based "static" refinement for single-block error paths
The first refinement that the predicate analysis does
is usually not interpolation-based, but uses some heuristics
for extracting predicates from the CFA.
When the error path consists of a single block,
this is useless, because we don't need real predicates.
We can use standard refinement instead,
which just consists of a single SAT check
(also no interpolation for single-block error paths).

Then we can do the heuristics-based refinement later
if we encounter a more interesting (longer) error path.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@44947 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c