Make ParallelAlgorithm generate nested algorithms from the threads that they run in
At least MathSAT5 does not like to be accessed from a thread that is different
from the thread it was created in
(see https://github.com/sosy-lab/java-smt/issues/339\). To work around this,
move the generation of nested algorithms into the individual threads that they run in.

Before this commit, the initialization of the algorithms nested in ParallelAlgorithm
was done sequentially at the construction of ParallelAlgorithm.
With this commit, this initialization moves to the run of ParallelAlgorithm,
and happens concurrently for the nested algorithms.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/svcomp24-parallel-portfolio-approach@45076 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c