Improve warning about spec to property mapping
This message gets emitted e.g. when running -valueAnalysis
with the default spec, which might be confusing for users.
"Will ignore it" does not state which implications this has,
i.e., can this warning be ignored or not. Actually we could
get rid of the warning, since there is no spc file for the
termination config, and the only reason this information
is currently used for is to determine whether there are
invariants allowed in a violation witness (because the
properties contain the termination property) or not.
But in the future this might change, so I would not remove
that warning, instead just clarify the warning text
and make the getSpecificationTypes method more restricted
and improve its JavaDoc to make clear to the person that
might use this in the future that this might miss some
specs that cannot be mapped to well-known property files.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@45082 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c