Remove the counterexample-check from the value analysis in SV-COMP24
Our sequential compositions in SV-COMP 2024 include
predicate-based analyses after the imprecise value analysis.
So instead of a counterexample check for value analysis,
just assume it is imprecise and continue with the more precise
analyses.
This saves us the time of the counterexample check (sometimes in the 100s)
and gives us no worse results because later analyses
find all of the alarms as well (in the SV-COMP setting)

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@45113 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c