Refactoring: move check for unsupported functions in predicate analysis
So far we check first for unsupported functions, and then for those that
we support. Let's swap this, such that if we add an implementation for a
function that happens to be covered by isUnsupportFunction() it still
works as it should.

This also combines better with CExpressionVisitorWithPointerAliasing:
All functions that are handled by that class are always checked first.
So it does not make sense to check for unsupported functions after those
functions handled by the latter visitor but before those handled by the
former visitor.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@45120 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c