Denylist all math.h functions in predicate analysis (if not supported)
Functions like expf() and logf() were so far treated as pure but
arbitrary functions. This is sound but imprecise.
We can guess that most code that calls these functions will actually not
be verifiable in a meaningful way without an implementation of these
functions, given the fact that we ignore assignments to irrelevant
variables completely. So let's better be more precise here.

This check is quite broad and denylists all functions in math.h,
but is only effective for those functions that are not explicitly
implemented in the predicate analysis.
So we do not loose any existing functionality.

For cases where we do want to overapproximate,
there is cpa.predicate.allowedUnsupportedFunctions
that can be used to allowlist functions.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@45121 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c