Improve soundness of ignoring void recursion
Enabling analysis.skipVoidRecursion will make CPAchecker ignore
recursive calls where the result of the function is not used.
This is unsound because it could still have side effect via passed
references (e.g. pass-by-pointer or even inside a struct
that is pass-by-value). This commit improves the situation by
only allowing void recursion to be ignored if the arguments
of the function call are of specific, known-to-be-safe CTypes.
I currently identified: CSimpleType, CEnumType, CBitFieldType;
This might not be maximal (e.g. an CElaborateType might be an enum)
but the worst case is we give up in cases where we are unsure
whether we are sound or not, and this allow-list can always be
extended. For other frontends/languages like Java the skipRecursion
option is still unsound, but with this changes it should be more
or less completely sound for C programs.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@45150 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c