Added option to export YAML witnesses directly from the counterexample
Currently we first produce the GraphML witness and then export the YAML witnesses from it.
Since the GraphML automaton abstracts from the ARG, this makes it very difficult to match the locations where the different waypoints should be in the original program.
To remedy this, now the YAML witnesses are directly exported from the counterexample.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@45152 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c