Revert "Fix time limit for SV-COMP 2024 config to 990s"
This reverts commit 8d6e9400a91466b38b16c019392bcbb9e1de61f1.

The rules for SV-COMP were always to use 15min.
It was considered to let tools run for a little bit longer
(ignoring the results produced after 15min),
but if we hard-code this in CPAchecker it causes lots of trouble,
for example when other users want to benchmark the SV-COMP configuration
of CPAchecker with the SV-COMP time limit of 15min.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@45156 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c