Enable memset etc. and the byte-level heap for non-interpolation configs
This is a part of #296 and #664 and a follow up to !107.
At least for all those cases where we use the predicate analysis
but do not use interpolation (e.g., BMC, k-induction, and counterexample
checks) it seems to work well to use the byte-level heap
and our recent memset/memcpy/etc. handling.

There are some fewer results, but of course this is much more precise
overall.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/664-predicate-enable-memset@45176 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c