Fix #1161: array access not treated as relevant like pointer access
The classification of our relevant variables treated
"p[0] = 1" different than "*p = 1" although they are exactly the same.
In the array case, p was not detected as relevant.
This can cause unsoundness and imprecision.

I cannot claim that I have a good understanding on why this change has
to look as it does. But I am just adding some more dependencies, which
should have no way to make the variable classification worse,
and the new code is based on the code that is used for classifying
pointer expressions in the same class.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@45178 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c