Fix #1165: RestartAlgorithm gives correct result if skipping last analysis
As explained in the issue, the RestartAlgorithm mistakenly returned
results from failed analysis runs as real analysis runs if the last
analysis was skipped (due to non-matching conditions like "if-recursive").
The reason for this was that RestartAlgorithm looks at
configFilesIterator to determine whether there is a next analysis,
but this does not take the analysis conditions into account.
Now we change that and let the iterator skip over all non-matching
analyses before we first look at it.

This has the disadvantage that the log messages for skipping the next
analysis now come before the log messages that describe the current's
analysis result. But all these log messages contain the analysis number,
so even if the order is wrong they can still be correctly interpreted,
and I see no easy way to fix the order
(we would need a "delayed logger" or so).

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@45210 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c