Improve handling empty lines and function return calls
In particular the reasoning for YAML violation witnesses is:
When a compound statement starts i.e. some statements in brackets {}, there may be an arbitrary amount of empty lines between the start of the compound statement and the next statement. This is now identified by looking at a possible range of line numbers instead of testing equality with a single line

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@45309 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c