Add aggressive caching to SwitchDetector
Reasons this caching is necessary:
* SwitchDetector is called for every occurrence of an assume edge in the ARG
* SwitchDetector traverses backwards from an assume edge of interest
  until it finds another assume edge or an if-statement.
  This can mean that SwitchDetector traverses many edges.

Combined, this produces a lot of overhead in WitnessFactory.
To avoid this, follow the strategy already performed with
WitnessFactory#loopEntryInfoMemo: Do not only cache the edges
that WitnessFactory asks about, but remember for any visited edge
whether it is part of a switch-case cascade or not. As soon as an edge
is not part of one, SwitchDetector can stop (or doesn't even have to be
started).

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/aggressive-caching-in-graphml-witness-export@45313 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c