Fix use of loopEntryInfoMemo in WitnessFactory
The memo was always filled with the original pEdge instead of
adding the currently considered node/edge.
Fix this so that WitnessFactory memorizes all edges that enter a loop
and does not recompute this unnecessarily.
Also add an early check whether information was already computed,
because this information may be requested often.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/aggressive-caching-in-graphml-witness-export@45314 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c