Fix handling of INT_MIN literals in CEX export
If we want to export a literal such as -9223372036854775808,
we actually need to make sure it shows up as -9223372036854775808UL
e.g. in the witness, because EclipseCDT currently cannot handle
-9223372036854775808L when reading in the witness.
C does not have negative integer constants, and 9223372036854775808L
is not a proper integer constant with a type (undefined behavior)!

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@45329 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c