In Test-Harness export, replace warnings with less grave messages
Replace log messages about issues with the hanress
with comments in the harness;
The messages have no influence on CPAchecker execution.
In one case, the harness may just not expose the property violation
in the way it is expected for CPA-witness2test. It may
still be exposed in some other, program-specific way.
In the other case, the harness creation failed.
This may happen often for some CPAs that produce imprecise counterexample
infos, so only show the information that harness-creation failed to the
user when they actually are interested in the harness (and look at it).

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/enable-test-harness-export-by-default@45369 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c