Rename remaining config/components/configselection-* files
All of them are SV-COMP specific, so they should get an appropriate name.
Unfortunately, these are used in several SV-COMP years,
so it is likely that they were changed over time
and old configs do not work anymore as intended.
But it is not worth investing effort to fix this.

The name of the renamed config files reflects the last year
in which they were used.
This way it is at least visible in the older configs
that a mismatch exists.


git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@45380 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c