Fix inclusion of SV-COMP'23 files in SV-COMP'22 configs
r42063 / 0f270f2e updated several files to refer to SV-COMP'23
even though these files were actually used for SV-COMP'22
and then ended up not being used for SV-COMP'23.
So we can easily revert these modifications
and keep the configs more consistent.

This does not change the actual content of the configs
because as it happens the svcomp22* files that are now included
as replacement for the svcomp23* files have the same content.


git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@45391 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c