Cleanup mess of duplicate SV-COMP'22 configs
When it was attempted to move the SV-COMP'22 configs to "unmaintained"
and creates copies with SV-COMP'23 in the name
in r41958 / 4b564f03, r41960 / 2b13bda0, r41961 / 4a6b6349, r41970 / 53d77100, and r41971 / 55a22d8f
the result was that all but two SV-COMP'22 config files
existed both in the "unmaintained" directory and outside.
Then r41972 / 65629aa9 removed only config/svcomp22.properties.
So the other files were still duplicated.

Several inconsistencies between the duplicates were then introduced,
but r45379 / e9c657c0, r45380 / 5d491ef9, r45384 / 6bb71763, and r45391 / a788edd7
cleaned this up.
So now we can simply remove the duplicates.

When looking at this commit with an SVN client,
you will see that all unmaintained/**/svcomp22* files are removed
and replaced by moving the files from outside "unmaintained" into the same location.
The diff that an SVN client shows, shows the changes
that were made to support the move (relative path changes)
and prove that the moved files behave the same as before.

When looking at this commit with a git client,
it will look as if simply all SV-COMP'22 files outside of "unmaintained"
were deleted (because git doesn't do proper rename tracking like SVN).
This proves that the remaining files behave the same as the files
inside "unmaintained" before.

The only other changes are necessary adjustments
to configs from before SV-COMP'22,
which unfortunately were changed to include SV-COMP'22 config files in the past.


git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@45393 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c