Fix #1164: restartAlgorithm.alwaysRestart was ignored for FALSE results
The option is documented as letting RestartAlgorithm always continue,
regardless of the previous analysis result.
But the implementation did this only if the analysis returned TRUE,
a result FALSE still let the RestartAlgorithm stop.
Now this is fixed and the option behaves as documented.

As discussed in #1164, for most existing users this does not matter.
For Lockator it is unclear, but the option was introduced (with its
documentation) specifically for Lockator, so we can assume that the
documented meaning is also intended for Lockator.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@45396 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c