Cleanup of valueAnalysis-*Cegar-weightedTraversal properties
These are included in subanalysis configs, but this shouldn't be the
case. So split them into a main config file and an includable config
file. For the -Cegar- case the config file basically consists only of
two lines and is used in one place, so don't bother with a new file,
just inline it.
Part of #838.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/838-fix-inclusion-of-main-configs@45409 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c