Make HarnessExporter define introduced dummy-functions
If the program under verification uses undefined function pointers,
the harness exporter initializes these with pointers to some
dummy functions of the correct type. Before this commit,
these dummy function were only declared, but not defined.
This meant that the code still did not compile,
because definitions were missing. We now introduce
dummy definitions to make this work.

To do this, make the auxiliary statements of TestValue
more flexible: Before, an auxiliary statement could only be
an actual C statement. We turn this into any auxiliary code,
so that it is possible to also include function definitions.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@45460 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c