Provide correct size of types to Eclipse CDT for parsing
We use CDT not only for parsing but also for related tasks
such as computing the types of specific expressions.
Due to the complex integer promotion rules of C,
these computations depend on the actual sizes of the types
on the respective platform
(certain expressions have different types on different platforms).
So CDT needs to know about the current platform,
and previously we did not provide this information.
Now we do, which means that we need to pass the machine model through.
Interesting that the missing information and incorrect type computation
was never noticed so far.

Fixes #1064, which is caused by such a type difference.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@45465 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c