Refactoring: replace use of MemoryLocation.parseExtendedQualifiedName()
This is a method that parses a name in the specific format returned by
MemoryLocation.toExtendedQualifiedName(), and should be used only in
specific circumstances.
The toString() output of types does not conform to this format,
and thus the method will always perform like forIdentifier(),
which is the correct method to use for
"just create a MemoryLocation for this arbitrary string".

Cf. r38080 / e8ad8a12.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@45474 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c