Fix for PointerCPA: ignore const and volatile
For the purpose of the PointerCPA, there should be no difference
whether somethings points to for example "const foo" or "foo"
(if "foo" is a type), as it may still alias.
Thus let's completely remove const and volatile before converting types
to strings.

This was actually a part of the problem for #1177
because even when the actual problem described there was fixed,
the result was still a wrong proof due to missing detection
of aliasing between a const and a non-const type.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@45476 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c