Fix #1177: Bug in field handling of PointerCPA
When reading a pointer value from a field, the PointerTransferRelation
failed to treat this in the same way as when reading from a simple
variable, and instead returned the location of the field.
Not only this inconsistency is fishy,
but the returned value was also just wrong
and now looks correct.

A more detailed description is in the issue.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@45477 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c