Remove duplicate code from ICCOperatorFactory
For most operators the strategy is to apply the operation to each
interval in the compound interval and then union the results.
We only need this code once.

This is basically a temporary commit that only serves to illustrate
an intermediate step in a bigger refactoring,
such that it is easier to see that each of the steps is correct.
The next commit will further refactor the code from this commit.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@45486 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c