Simplify operations on two CompoundBitVectorInterval instances
Some of these operations are simply performed on every combination
of intervals in both operands. This is currently spread over several
classes, with the two loops being implemented in different classes.
It is much easier to see what happens if we simply write down
the nested for loop explicitly.
We can also share the code across all operations instead of each having
its own implementation.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@45487 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c