Simplify remaining operations on two CompoundBitVectorInterval instances
For add and multiply a different implementation strategy had been chosen
than for the other operators
(which were refactored in the previous two commits),
because these two operators are commutative.
However, in the end the same happened:
every interval from the first operand was combined with every operand
from the second operand, just even more difficult to see
because the same method and for loop was used for both iterations
with the order of operands swapped in between.

Now that we have the nice helper method that does this explicitly,
we can also use this for add and multiply.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@45488 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c