Fix hashCode() of SMGPointsToEdge
Calling super.hashCode() means that the hashcode differs for every
object, which is obviously wrong and violates the contract.
Looking at the equals() method it becomes clear that the hashcode of the
target specifier should be used.

It seems that r41571 / 6230f4e8 attempted to fix this,
but erroneously kept the super.hashCode() call.

Issue found by Google Error Prone 2.24.0.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@45509 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c