Abstract states are not required to match if the transition in the automaton should be taken
If this condition is left as is, the transition is only taken during the strengthening, since that is the moment at which AbstractStates are available. The Problem is that we want to use the assumptions present in the automaton transition to strengthen other AbstractStates, for example the value analysis state, we require the match to be done at the moment of the TransferRelation call, at which point no AbstractStates exist when asking for successors of the transition.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/direct-witness-export@45599 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c