Improve naming convention
Currently YAML witnesses stand for all witnesses which are exported or read from a YAML file, where the specific version cannot be determined, if there is more than a single one. Currently this is only the case for the export of correctness witnesses.

If the Version is known the version is used.

Ig the version is clear from context (default is version 2), then no further identifier is required

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/direct-witness-export@45693 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c