Make symbolic execution use concrete values from equality assumptions
Use the existing value-analysis optimization
instead of introducing new symbolic values.
This can lead to a lot faster analyses in cases like this:

```
int x;
while (x = __VERIFIER_nondet_int()) {
  // .. something that doesn't change values ..
}
```

Basic symbolic execution introduces a new symbolic value for every
assignment to `x`. The value-analysis optimization
that is now also possible for symbolic execution
assigns value '1' to x on each call, as long as x is only
used in this boolean context.
This leads to a fast stop.

In this process, we remove a rather strict checkState
that made sure that no symbolic value is replaced with
a concrete value.
This check was added to ensure that concretization happens
through the constraints CPA.
While conceptually a nice idea, this is not a requirement
and not feasible for this optimization.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@45958 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c