Use list instead of object array
Object arrays are in general discouraged (cf. style guide)
and should be used only where necessary.
The current place is clearly not such a situation
because even if callers of
buildConcreteListWithDifferentPtrTargetOffsetsInEndAndBeginning
would have really need an array,
they could just do the array creation themselves.
But none of the callers needs an array.

Changing this also avoids a problem with our Refaster rules
that produce a patch here that differs only in formatting.

There are many more array uses in this class,
and likely all of them should be eliminated.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@46454 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c