Rename some DSS methods to make them more expressive
A method that returns the boolean decision whether to end an analysis
should not be named 'end', but 'shouldEnd'. Otherwise it sounds
as if the method actually ends something. Also modify the method to
match the positive case that is checked ('shouldProceed').

Also clarify some other method names.
For example the ProceedOperator processes an abstract state
to decide whether to proceed with analysis, but it does not
proceed itself.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/abstraction-with-block-summaries@46518 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c