Remove argument check in JClassType that is too strict
The check makes sure that no subclass is registered with its superclass twice.
If this happens it may be a bug, but it does not hurt because
the collection of subclasses is a Set.
But this check triggers if a single CPAchecker instance is used
to perform two runs, as may also happen in unit tests.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@46549 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c