Provide logger to DSS workers through dependency injection
It's an unnecessarily complicated architecture that subclasses of 'BlockSummaryWorker'
should access the parent's logger through a protected method when they can just
store and manage a logger for themselves.
In addition, the BlockSummaryWorker should not create
something of orthogonal concern like a logger on creation,
but get it from the outside.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/abstraction-with-block-summaries@46600 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c