The information about what variables are in scope is currently only tracked for C programs, therefore if this information is not available we do not filter for the variable names
These functions in which these checks are performed are only used when
exporting witnesses. For witnesses in version 2.0 the checks should stay
there, since they actually allow us to find error and the witnesses are
only defined for C programs. Currently the checks are only necessary for
exporting witnesses in version 1.0.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/track-variables-in-scope@46668 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c