Strengthen types in Witness class
It is best practice to always use the Immutable* types
for fields that only contain instances of these types,
such that one can immediately see that this class is immutable.
It is also best practice to use strong types as return types
(like ListMultimap instead of Multimap, Immutable*)
such that callers can rely on the guarantees of these types.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@46711 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c