Move method generateLocationString into utility class
This method is inside SLARGToDotWriter but also used for witness export.
This means that our witness export could easily change
if someone changes SLARGToDotWriter, which we do not want.
Utility methods from component A should not be used in component B,
they should be moved to central classes.

Furthermore, the method is much more general than the current usage.
The only change made in this commit is the names and the JavaDoc
of the method.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@46712 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c