Fix bug in convertIntegerRangesToStringCollapsed
If the values passed to this method would be such that there was a range
of exactly 3 int values at the end (e.g., "1,5-7"),
the method would skip the middle value
(the result of the example was "1,5,7").

Furthermore, switch from having two state variables
of the finite automaton (state and separator)
to a single variable, which is much easier to understand.

As the bug shows it is important to have tests for such methods,
thus add some.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@46714 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c