Fix a bug when deserializing correctness witnesses
There was a bug when deserializing the classes due to how jackson manages the deserialization of abstract classes and the delegation to other classes. Fixing this required a manual delegation due to the nested nature of invariants.

In the process of fixing this, I also refactored the name of the class representing an abstract invariant i.e. loop_invariant, location_invariant or function_contract. This could be done due to the proposed format using function contracts changing in the last weeks and therefore making a unification of the interfaces easier.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@46736 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c