Refactor interface by adding all methods which are required to export correct witnesses with function contracts
Since not all functions are required to return every type of witness, IMO it makes sense for them to throw an NotImplementedError if they are not needed. In particular considering that the interface may need to be expanded. The idea behind having four different functions and not one with multiple flags is such that the flexibility on how to implement them is on the AbstractState and it is not forced to somehow deal with all the possible options.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@46738 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c