PredicateAbstraction: collect all atoms when splitting formulas into inequalities, not just the first atom.
This issue was found in a test.
A new solver used "a=1" instead of "1=a", due to solver-internal formula-rewriting,
and the set of atoms after splitting atoms into inequalities missed "1<=a" and provided "a<=1".
Overall, only the first of the inqualities was ever considered for further processing.
It is unclear whether this was a bug or not.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/update-javasmt-4.1.1@46757 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c