Provide container images on Docker Hub as well
In order to make it more convenient to use the container images of
CPAchecker, we do not only want to provide them with the long name
registry.gitlab.com/sosy-lab/software/cpachecker,
but also as docker.io/sosylab/cpachecker,
which in most cases can even be abbreviated to sosylab/cpachecker.

This adds the respective CI jobs and updates the documentation.

The images will also stay on GitLab for the foreseeable future.
The internal images used by CI will continue to exist only on GitLab.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@46848 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c