Move compiled .class files in CPAchecker project from bin/ to classes/
This is a preparation for the creation of bin/cpachecker
as part of #632.

Everything should continue to work as before,
just that when switching between commits before and after this change
an incremental build run has to compile more classes.

Old files in bin/ are not automatically cleaned up and remain,
but can be deleted manually.


git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/632-new-cpachecker-executable@46862 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c