Set locale in Docker images to UTF-8
Previously cpachecker executions printed this warning,
which may be confusing to users:

JVM property file.encoding is set to non-standard value 'ANSI_X3.4-1968'. This is not recommended and output files might be written in unexpected encodings. (CPAMain.main, WARNING)

Since we don't want to confuse them when using docker, this commit
gets rid of it.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/use-utf8-in-docker-images@46871 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c