Fix startup script of CPAchecker for Windows in error cases
Calling EXIT in a batch file does not terminate only the batch script,
but also the cmd.exe process that called the script
(usually the user's command-line session)
and potentially even the terminal window.
(This is similar to what happens on Linux when exit is called
in a *sourced* shell script.)
Of course we do not want this, and "EXIT /B" is the solution
that only terminates the batch script.
On regular exit (end of script) everything is working already
as it should.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@46873 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c