Use openjdk-17-jre in docker images
Instead of openjdk-17-jre-headless.
Some awt libraries are required for CPAchecker's svg exports.
Somehow this _currently_ is not an issue on our Dockerfile.release
(based on Ubuntu 20), but only on Dockerfile.latest (based on latest Ubuntu).
This commit still changes both to not encounter the same issue
in the future.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/make-svg-output-work-in-docker-dev-image@46880 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c