Make ARG/CFA export in SVG gracefully handle missing awt libraries
CPAchecker's SVG export requires access to the X11 font manager.
If that can not be loaded because of missing libraries
or no fonts, catch the resulting internal error
and log it gracefully.

As part of this commit, revert the changes from commit efb0f69c472cc5ed85d358aa720eabf48f94917c.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/make-svg-output-work-in-docker-dev-image@46889 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c