Script and definitions for creating a .deb package for CPAchecker
This is the base for creating a usable Debian/Ubuntu package,
with the minimal amount of work.

The package is created from a release ZIP of CPAchecker,
so both the source and the binary package
contain a fully built CPAchecker.

CPAchecker is installed into /opt/cpachecker
because this is significantly easier than a proper package layout.

Part of #1220.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/1220-deb-package@46900 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c