Add new --long-form command-line arguments to CPAchecker
Part of #1218, and the concrete argument names are discussed in
https://gitlab.com/sosy-lab/software/cpachecker/-/issues/1218#note_1950486702

So far this adds only new arguments, and does not mark the old arguments
as deprecated. This includes that a mixture of old and new options
is accepted right now, which we actually do not want.
New short arguments like "-c" are also not yet added
and neither is support for calling an analysis like "--default".

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/1218-new-command-line-arguments@46916 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c