Refactor comments in benchmark definitions
Several benchmark definitions have parts that are commented out.
Unfortunately, "--" is a forbidden string inside XML comments,
but in the next commits we want to update these benchmark definitions
to use "--long-form" arguments, which would produce invalid syntax
when doing the replacement inside comments.
So we change those comments where this is relevant to something else
that is not a comment but also ignored by XML parsers:
an undefined processing instruction.
This seems to work fine with BenchExec and xmlstarlet does not complain
about it.

Solution is taken from https://stackoverflow.com/a/14650451/396730

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/1218-new-command-line-arguments@46932 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c