Print warning for deprecated command-line arguments and remove from docs
This affects those five arguments for which we decided
to not provide a new "--long-form" replacement
because they are rarely used.

Because replacing them requires more than just replacing them on the
command line we print a separate warning
and do not let this be part of the general warning about
old-style command-line arguments.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/1218-new-command-line-arguments@46948 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c