Remove some scripts from scripts/ directory
report-generator.py is not necessary anymore and a noop for many years.

runexecutor.py is just the same as BenchExec's runexec.
We provide an extended copy of BenchExec here,
so we have benchmark.py and for convenience also table-generator.py,
but those who want to use runexec should simply install BenchExec.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@46973 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c