Fix handling of arguments like "--default" in VerifierCloud WebClient mode
Runs submitted via the VerifierCloud WebClient currently still always
use the old-style command-line arguments for the actual CPAchecker
execution. This works because when uploading the run, all arguments
are turned into options, and then the WebClient creates a fresh command
line from the options.
However, for options that specify an analysis we need to make sure
that it is always normalized to its bare form (without dashes).
This was missing.

Cf.
https://gitlab.com/sosy-lab/software/verifiercloud/-/issues/685#note_1973172115

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@46977 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c