Error out for analysis without spec if $specification is in config
Some config files use $specification in the list of CPAs
in order to define where to insert the specification automata
in the tree of CPAs.
So far, if no specification automata where provided,
nothing would be inserted in this place.
No we define that $specification requires at least one specification
automaton, and produce an error if it is missing.

This is actually good for --ltl,
which makes sense only if an LTL specification is given.
Apart from this the change should only affect custom configs,
because all our provided configs should include the default spec anyway.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@47028 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c