Add log message when CPAchecker is using default specification
We considered requiring the user to always explicitly specifying
the specification, but decided that it is enough to inform them
in a log message about the semantics of the used specification.

It is not so nice to hard-code this inside the CPAchecker class
with the path etc., but it is good for now.
A more general solution would be to define some description
in every specification file and print this out,
but the current syntax does not allow this.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@47029 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c