Optimize our container images by deleting APT package lists
According to
https://docs.docker.com/build/building/best-practices/#apt-get,
"apt-get clean" is executed automatically by Ubuntu images,
but this does not delete the package lists, which take up ~50 MB.
So deleting them manually is recommended.

For the images where we run apt-get twice to install Google Chrome,
the best way is to pack everything into one RUN command.
Otherwise we would have to call apt-get update twice,
because due to the layer caching one can not rely on APT package lists
from a previous RUN command (this was a bug in the Dockerfiles so far).

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@47081 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c